Added layout for cms page create, ... * Implemented CMS#Preferences#AuthorVisible in CMS Template * Implemented CMS#Preferences#UpdatersVisible in CMS Template
