+++ internal/whgoxy/http/cms/update.go
@@ -5,8 +5,12 @@ import (
 )
 
 type CMSPageUpdate struct {
-	UpdatedAt     time.Time `bson:"updated_at" json:"updated_at"`           // Time of update
-	UpdaterUserID string    `bson:"updater_user_id" json:"updater_user_id"` // UserID of updater, -1 = System
+	// Time of update
+	UpdatedAt time.Time `bson:"updated_at" json:"updated_at"`
+	// UserID of updater, -1 = System
+	UpdaterUserID string `bson:"updater_user_id" json:"updater_user_id"`
+	// Patch of content
+	Patch string `bson:"patch" json:"patch"`
 }
 
 func (p *CMSPage) GetLastUpdate() (res *CMSPageUpdate) {
